feat: add IRQReader abstraction
This abstraction is intended for proper IRQ handling sequence.

Multiple IRQ reader implementations can be present as the hardware might have different requirements.

When an interrupt is detected an element is pushed to the queue.

The queue is processed by the IRQ handler thread.

The event also have the following informations:
 - `device`: what device has the triggered IRQ
- `timestamp`: IRQ timestamp, as the IRQ handler is asynk we need to keep track of original time
 - `source`: the name of the IRQ reader
 - `data`: additional payload that the IRQ reader want to include in the queue, this is intended for the global IRQ0 register flag to be passed.