feat: add HookPulseDetector state machine for pulse phones
Some older phones uses a series of pulses to send the digits.
Modern phones may have an "R" button that simulate the
 older manual "Flash hook" signal.
This state machine keep tracks of the sequence of line changes
 and decode them  to the proper phone digit or hook signal
 using the IRQ timestamps.

The idea is to foward the digits or the ONHOOK/OFFHOOK
 timeout events to the VoiceChannel to be used for dial or
 to terminate a call.

The timings are configurable by a series of values in the `fxs` node.
Below an example that should be compatible with ETSI/Italian phones:

config fxs 'Phone1'
	option audio_slot '0'
	option impedance 'TBR21'
	option ring_pattern '60(2/4)'
	option tone_busy '425@-5;20(.5/.5/1)'
	option tone_dial '425@-5;10(.2/.2/1,.6/1/1)'
	# Telecom POTS documentation chapter 5.8.1
	option min_hook_timeout '0.350'
	option min_digit '0.030'
	option max_digit '0.075'
	option min_flash '0.080'
	option max_flash '0.100'
	option min_inter_digit '0.250'

As of now this class only print the detected events.